ai-book-bot/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
pydantic==2.6.4
python-dotenv==1.0.1
aiohttp==3.9.5
alembic==1.13.1
aiosqlite==0.19.0   # <--- добавить сюда
